renamed principal assignments to assignment group
