+++ sso/assignment-group/.npmrc
