+++ sso/assignment-group/apis/sso-admin-2020-07-20.min.json
