+++ sso/assignment-group/package.json
@@ -1,7 +1,7 @@
 {
-    "name": "community-sso-principalassignments",
-    "version": "0.1.0",
-    "description": "AWS custom resource provider named Community::SSO::PrincipalAssignments.",
+    "name": "community-sso-assignmentgroup",
+    "version": "0.1.2",
+    "description": "AWS custom resource provider named Community::SSO::AssignmentGroup.",
     "main": "dist/handlers.js",
     "files": [
         "dist",
