+++ sso/assignment-group/resource-role.yaml
