+++ sso/assignment-group/src/models.ts
@@ -6,7 +6,7 @@ export class ResourceModel extends BaseModel {
     ['constructor']: typeof ResourceModel;
 
     @Exclude()
-    public static readonly TYPE_NAME: string = 'Community::SSO::PrincipalAssignments';
+    public static readonly TYPE_NAME: string = 'Community::SSO::AssignmentGroup';
 
     @Exclude()
     protected readonly IDENTIFIER_KEY_RESOURCEID: string = '/properties/ResourceId';
