+++ sso/assignment-group/template.yml
@@ -1,6 +1,6 @@
 AWSTemplateFormatVersion: "2010-09-09"
 Transform: AWS::Serverless-2016-10-31
-Description: AWS SAM template for the Community::SSO::PrincipalAssignments resource type
+Description: AWS SAM template for the Community::SSO::AssignmentGroup resource type
 
 Globals:
   Function:
