+++ sso/assignment-group/tsconfig.json
