--- sso/principal-assignments/example.yml
@@ -1,37 +1,0 @@
- 
-AWSTemplateFormatVersion: 2010-09-09
-
-Parameters:
-  instanceArn: 
-    Type: String
-    Default: arn:aws:sso:::instance/ssoins-680434798a99c2fa
-
-  accountId: 
-    Type: String
-    Default: '102625093955'
-
-Resources:
-
-  AdminPS:
-    Type: AWS::SSO::PermissionSet
-    Properties: 
-      Name: Admin
-      Description: Full access to AWS services allowed
-      InstanceArn: !Ref instanceArn
-      ManagedPolicies: 
-        - arn:aws:iam::aws:policy/AdministratorAccess
-      SessionDuration: PT12H
-
-
-  TestPrincipalAssignments:
-    Type: Community::SSO::PrincipalAssignments
-    Properties:
-      InstanceArn: !Ref instanceArn
-      PermissionSets:
-      - !Ref AdminPS
-      PrincipalId: 93670582cf-0e506d8f-2542-4048-b551-74f5ef7497e5
-      PrincipalType: GROUP
-      Targets:
-      - TargetType: AWS_ACCOUNT
-        TargetIds:
-        - !Ref accountId
